VolumeChanged with int param.

+++ src/main/java/com/shuffle/scplayer/core/PlayerListener.java
@@ -27,5 +27,5 @@ public interface PlayerListener {
 
     void onTokenLost();
 
-    void onVolumeChanged(short volume);
+    void onVolumeChanged(int volume);
 }
